requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
deepeval>=0.20.9
langchain>=0.0.267
langchain-openai>=0.0.2
pydantic>=2.4.0
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.1
nest-asyncio>=1.5.6
